+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- New method `get_data` for `DataFile` provides a simple way to extract data selectively from data files.
 - Added dependency on xmake `openblas` package to support calls to BLAS/LAPACK routines.
 - New exception `QuantityError` and the associated procedure `RaiseQuantityError` for errors related to the processing of data/quantities.
 - New module `vibdata` to store vibrational data extracted from input files.